requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.7
defusedxml==0.7.1
Django==6.0.4
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-oauth-toolkit==3.2.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-silk==5.5.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gprof2dot==2025.4.14
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.7
kombu==5.6.2
oauth==1.0.1
oauthlib==3.3.1
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyGithub==2.9.1
Pygments==2.20.0
PyJWT==2.12.1
PyNaCl==1.6.2
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-social-auth==0.3.6
python3-openid==3.2.0
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
